Prepare removal of alpha SDK
