+++ src/cpprealm/app.hpp
@@ -20,7 +20,7 @@
 #define CPP_REALM_APP_HPP
 
 #include <cpprealm/task.hpp>
-#include <cpprealm/db.hpp>
+#include <cpprealm/experimental/db.hpp>
 #include <cpprealm/internal/bridge/sync_session.hpp>
 #include <cpprealm/internal/bridge/sync_error.hpp>
 #include <cpprealm/internal/bridge/sync_manager.hpp>
