+++ src/cpprealm/experimental/accessors.hpp
@@ -257,10 +257,10 @@ namespace realm::experimental {
                 }
                 auto table = obj.get_target_table(key);
                 internal::bridge::obj m_obj;
-                if constexpr (managed<T>::schema.HasPrimaryKeyProperty) {
-                    auto pk = (*lnk).*(managed<T>::schema.primary_key().ptr);
+                if constexpr (managed<T, void>::schema.HasPrimaryKeyProperty) {
+                    auto pk = (*lnk).*(managed<T, void>::schema.primary_key().ptr);
                     m_obj = table.create_object_with_primary_key(realm::internal::bridge::mixed(serialize(pk.value)));
-                } else if (managed<T>::schema.is_embedded_experimental()) {
+                } else if (managed<T, void>::schema.is_embedded_experimental()) {
                     m_obj = list.create_and_insert_linked_object(i);
                 } else {
                     m_obj = table.create_object();
@@ -270,7 +270,7 @@ namespace realm::experimental {
                              m_obj, m_obj.get_table().get_column_key(p.name),
                              (*lnk).*(std::decay_t<decltype(p)>::ptr)), ...);
                 }, managed<T, void>::schema.ps);
-                if (!managed<T>::schema.is_embedded_experimental()) {
+                if (!managed<T, void>::schema.is_embedded_experimental()) {
                     list.add(m_obj.get_key());
                 }
             }
@@ -318,8 +318,8 @@ namespace realm::experimental {
             for (auto& [k, v] : value) {
                 if (v) {
                     internal::bridge::obj m_obj;
-                    if constexpr (managed<T>::schema.HasPrimaryKeyProperty) {
-                        auto pk = (*v).*(managed<T>::schema.primary_key().ptr);
+                    if constexpr (managed<T, void>::schema.HasPrimaryKeyProperty) {
+                        auto pk = (*v).*(managed<T, void>::schema.primary_key().ptr);
                         m_obj = d.create_and_insert_linked_object(k, pk.value);
                     } else {
                         m_obj = d.create_and_insert_linked_object(k);
@@ -348,11 +348,11 @@ namespace realm::experimental {
             }
             auto table = obj.get_target_table(key);
             internal::bridge::obj m_obj;
-            if constexpr (managed<T>::schema.HasPrimaryKeyProperty) {
-                auto pk = (*value).*(managed<T>::schema.primary_key().ptr);
+            if constexpr (managed<T, void>::schema.HasPrimaryKeyProperty) {
+                auto pk = (*value).*(managed<T, void>::schema.primary_key().ptr);
                 m_obj = table.create_object_with_primary_key(realm::internal::bridge::mixed(serialize(pk.value)));
                 obj.set(key, m_obj.get_key());
-            } else if (managed<T>::schema.is_embedded_experimental()) {
+            } else if (managed<T, void>::schema.is_embedded_experimental()) {
                 m_obj = obj.create_and_set_linked_object(key);
             } else {
                 m_obj = table.create_object();
