+++ src/cpprealm/experimental/managed_objectid.cpp
@@ -1,5 +1,5 @@
 #include <cpprealm/experimental/managed_objectid.hpp>
-#include <cpprealm/persisted.hpp>
+#include <cpprealm/rbool.hpp>
 
 namespace realm::experimental {
     __cpprealm_build_experimental_query(==, equal, realm::object_id);
