+++ src/cpprealm/experimental/managed_uuid.cpp
@@ -1,5 +1,5 @@
 #include <cpprealm/experimental/managed_uuid.hpp>
-#include <cpprealm/persisted.hpp>
+#include <cpprealm/rbool.hpp>
 
 namespace realm::experimental {
     __cpprealm_build_experimental_query(==, equal, realm::uuid)
