+++ src/cpprealm/persisted.hpp
@@ -19,48 +19,48 @@
 #ifndef realm_persisted_hpp
 #define realm_persisted_hpp
 
-#include <cpprealm/internal/type_info.hpp>
-#include <cpprealm/internal/bridge/object.hpp>
-#include <cpprealm/internal/bridge/obj.hpp>
-#include <cpprealm/rbool.hpp>
+//#include <cpprealm/internal/type_info.hpp>
+//#include <cpprealm/internal/bridge/object.hpp>
+//#include <cpprealm/internal/bridge/obj.hpp>
+//#include <cpprealm/rbool.hpp>
+//
+//namespace realm::experimental {
+//    template<typename>
+//    struct box_base;
+//}
 
-namespace realm::experimental {
-    template<typename>
-    struct box_base;
-}
-
-namespace realm {
-    namespace cpprealm {
-        constexpr size_t npos = size_t(-1);
-    }
-    namespace {
-        template<typename T>
-        using is_optional = internal::type_info::is_optional<T>;
-    }
-    namespace schemagen {
-        template <typename Class, typename ...Properties>
-        struct schema;
-        template <auto Ptr, bool IsPrimaryKey>
-        struct property;
-    }
+//namespace realm {
+//    namespace cpprealm {
+//        constexpr size_t npos = size_t(-1);
+//    }
+//    namespace {
+//        template<typename T>
+//        using is_optional = internal::type_info::is_optional<T>;
+//    }
+//    namespace schemagen {
+//        template <typename Class, typename ...Properties>
+//        struct schema;
+//        template <auto Ptr, bool IsPrimaryKey>
+//        struct property;
+//    }
 
 //// MARK: Equatable
 
-inline rbool operator &&(const rbool& lhs, const rbool& rhs) {
-    if (lhs.is_for_queries) {
-        lhs.q.and_query(rhs.q);
-        return lhs;
-    }
-    return lhs.b && rhs.b;
-}
-inline rbool operator ||(const rbool& lhs, const rbool& rhs) {
-    if (lhs.is_for_queries) {
-        lhs.q = lhs.q || rhs.q;
-        return lhs;
-    }
-    return lhs.b && rhs.b;
-}
+//inline rbool operator &&(const rbool& lhs, const rbool& rhs) {
+//    if (lhs.is_for_queries) {
+//        lhs.q.and_query(rhs.q);
+//        return lhs;
+//    }
+//    return lhs.b && rhs.b;
+//}
+//inline rbool operator ||(const rbool& lhs, const rbool& rhs) {
+//    if (lhs.is_for_queries) {
+//        lhs.q = lhs.q || rhs.q;
+//        return lhs;
+//    }
+//    return lhs.b && rhs.b;
+//}
 
-}
+//}
 
 #endif /* realm_persisted_hpp */
